修正＆追加しました
ゲーム→リザルト
リザルト→タイトル
へのループを作りました。
Optionウィンドウを閉じられるようにしました
