+++ Assets/GURA/Prefab/FadeIn.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5fae60c40227cd642b48fd7ae64c6989
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
