+++ Assets/GURA/Prefab/FadeOut.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ffdde12cae1249440bef31fcaf867300
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
