+++ Assets/GURA/Scenes/ResultScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6b99017d7cef28a4298c24bff4c0c3e8
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
