+++ Assets/GURA/Script/ResultDirector.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: aed3a5f0cbea6f24b8d2ba6ea7fc69d9
+guid: 728e0cc672482ce4f96aa9e28018b960
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
