+++ Assets/GURA/Script/TitleDirector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 00c3735b1891a0641ba5dc24f0df9548
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
