+++ ProjectSettings/EditorBuildSettings.asset
@@ -11,4 +11,7 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/GURA/Scenes/GameScene.unity
     guid: 78b27228128e96a43bd4f49f7034c6f3
+  - enabled: 1
+    path: Assets/GURA/Scenes/ResultScene.unity
+    guid: 6b99017d7cef28a4298c24bff4c0c3e8
   m_configObjects: {}
