+++ UserSettings/EditorUserSettings.asset
@@ -11,6 +11,12 @@ EditorUserSettings:
     RecentlyUsedScenePath-1:
       value: 224247031146467c383c2d6d23205a1e1304571e253c2a351e2a183de7ae2136ebf32f
       flags: 0
+    RecentlyUsedScenePath-2:
+      value: 224247031146467c383c2d6d23205a1e1304570d2d2523032e2c1336acf53a31f6fe
+      flags: 0
+    RecentlyUsedScenePath-3:
+      value: 224247031146467c383c2d6d23205a1e13045718293b333c391a1e36ece57a2decee22f0
+      flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
       flags: 0
